+++ Tests/CleanArchitecture.Comprehensive/CleanArchitecture.Comprehensive.Application/Customers/GetCustomersPaged/GetCustomersPagedQueryHandler.cs
@@ -1,0 +1,37 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using AutoMapper;
+using CleanArchitecture.Comprehensive.Application.Common.Pagination;
+using CleanArchitecture.Comprehensive.Domain.Repositories.DDD;
+using Intent.RoslynWeaver.Attributes;
+using MediatR;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.QueryHandler", Version = "1.0")]
+
+namespace CleanArchitecture.Comprehensive.Application.Customers.GetCustomersPaged
+{
+    [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
+    public class GetCustomersPagedQueryHandler : IRequestHandler<GetCustomersPagedQuery, PagedResult<CustomerDto>>
+    {
+        private readonly ICustomerRepository _customerRepository;
+        private readonly IMapper _mapper;
+
+        [IntentManaged(Mode.Merge)]
+        public GetCustomersPagedQueryHandler(ICustomerRepository customerRepository, IMapper mapper)
+        {
+            _customerRepository = customerRepository;
+            _mapper = mapper;
+        }
+
+        [IntentManaged(Mode.Fully, Body = Mode.Fully)]
+        public async Task<PagedResult<CustomerDto>> Handle(
+            GetCustomersPagedQuery request,
+            CancellationToken cancellationToken)
+        {
+            var customers = await _customerRepository.FindAllAsync(request.PageNo, request.PageSize, cancellationToken);
+            return customers.MapToPagedResult(x => x.MapToCustomerDto(_mapper));
+        }
+    }
+}