+++ Tests/CleanArchitecture.Comprehensive/CleanArchitecture.Comprehensive.Application/Customers/GetCustomersPaged/GetCustomersPagedQueryValidator.cs
@@ -1,0 +1,24 @@
+using FluentValidation;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.Application.MediatR.FluentValidation.QueryValidator", Version = "2.0")]
+
+namespace CleanArchitecture.Comprehensive.Application.Customers.GetCustomersPaged
+{
+    [IntentManaged(Mode.Fully, Body = Mode.Merge)]
+    public class GetCustomersPagedQueryValidator : AbstractValidator<GetCustomersPagedQuery>
+    {
+        [IntentManaged(Mode.Merge)]
+        public GetCustomersPagedQueryValidator()
+        {
+            ConfigureValidationRules();
+        }
+
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Performance", "CA1822:Mark members as static", Justification = "Depends on user code")]
+        private void ConfigureValidationRules()
+        {
+            // Implement custom validation logic here if required
+        }
+    }
+}